docs(measure): define advisory trust boundary
