+++ crates/measure/src/event.rs
@@ -148,46 +148,3 @@ pub enum ApplyOutcome {
     /// capacity or evicting authenticated credit.
     IgnoredUnknownPeer,
 }
-
-/// Complete result of applying an observation to a bounded ledger.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ApplyReport<P> {
-    outcome: ApplyOutcome,
-    evicted_peer: Option<P>,
-}
-
-impl<P> ApplyReport<P> {
-    /// Construct a report for an applied transition and its optional eviction.
-    pub(crate) const fn applied(evicted_peer: Option<P>) -> Self {
-        Self {
-            outcome: ApplyOutcome::Applied,
-            evicted_peer,
-        }
-    }
-
-    /// Construct a report for an observation without attributable identity proof.
-    pub(crate) const fn ignored_unattributable() -> Self {
-        Self {
-            outcome: ApplyOutcome::IgnoredUnattributable,
-            evicted_peer: None,
-        }
-    }
-
-    /// Construct a report for local evidence about an unknown peer.
-    pub(crate) const fn ignored_unknown_peer() -> Self {
-        Self {
-            outcome: ApplyOutcome::IgnoredUnknownPeer,
-            evicted_peer: None,
-        }
-    }
-
-    /// Whether the requested observation changed retained state.
-    pub const fn outcome(&self) -> ApplyOutcome {
-        self.outcome
-    }
-
-    /// Peer deterministically removed to admit this transition, if any.
-    pub const fn evicted_peer(&self) -> Option<&P> {
-        self.evicted_peer.as_ref()
-    }
-}
