+++ crates/measure/src/lib.rs
@@ -28,11 +28,11 @@ pub use credit::CreditScore;
 pub use error::MeasureError;
 pub use error::PolicyError;
 pub use event::ApplyOutcome;
-pub use event::ApplyReport;
 pub use event::Authentication;
 pub use event::MeasurementBatch;
 pub use event::MeasurementEvent;
 pub use event::Metric;
+pub use ledger::ApplyReport;
 pub use ledger::LedgerReconciliation;
 pub use ledger::MeasurementLedger;
 pub use ledger::MeasurementPage;
